Restructure Serial/StationaryMappingWithSolver
