+++ tests/serial/stationary-mapping-with-solver-mesh/StationaryMappingWithSolverMesh2D.xml
