+++ tests/serial/stationary-mapping-with-solver-mesh/StationaryMappingWithSolverMesh3D.xml
