+++ tests/tests.cmake
@@ -92,6 +92,10 @@ target_sources(testprecice
     tests/serial/multi-coupling/helpers.hpp
     tests/serial/multiple-mappings/MultipleFromMappings.cpp
     tests/serial/multiple-mappings/MultipleToMappings.cpp
+    tests/serial/stationary-mapping-with-solver-mesh/StationaryMappingWithSolverMesh2D.cpp
+    tests/serial/stationary-mapping-with-solver-mesh/StationaryMappingWithSolverMesh3D.cpp
+    tests/serial/stationary-mapping-with-solver-mesh/helpers.cpp
+    tests/serial/stationary-mapping-with-solver-mesh/helpers.hpp
     tests/serial/three-solvers/ThreeSolversExplicitExplicit.cpp
     tests/serial/three-solvers/ThreeSolversExplicitImplicit.cpp
     tests/serial/three-solvers/ThreeSolversImplicitExplicit.cpp
